+++ routes/api/endpoints/results.js
@@ -1,0 +1,7 @@
+const router = require('express').Router();
+
+router.post('/',async function (req,res) {
+    res.send('This is the POST results endpoint')
+})
+
+module.exports = router;
