+++ routes/api/router.js
@@ -1,5 +1,6 @@
 // Load the routers for each resource of the api
-const testQuestionsRouter = require('./api/test-questions');
+const testQuestionsRouter = require('./endpoints/test-questions');
+const resultsRouter       =  require('./endpoints/results');
 
 const { Router } = require("express");
 
@@ -10,5 +11,6 @@ const router = Router();
 
 // Mount the routes here
 router.use('/test-questions',testQuestionsRouter);
+router.use('/results',resultsRouter);
 
 module.exports = router;
